refactor(utils): remove dead exports and tighten the utility split
Follow-up to the utilityObsidian decomposition, addressing adversarial-review
findings:

- Delete dead exports getChoiceType and getNaturalLanguageDates (zero callers
  anywhere; getChoiceType's generic guard signature was also unsound).
- Extract the pure markdown-link helpers (convertLinkToEmbed,
  extractMarkdownLinkTarget) into utils/markdownLinks.ts and export them
  normally, removing editorInsertion's `__test` back-door. The test now
  imports them directly.
- Move excludeKeys off the Obsidian-facing barrel: choiceService (its only
  caller) imports utils/excludeKeys directly.

Behavior-preserving: tsc/lint/build clean, 2127 tests green, plugin loads with
no errors, date+template run verified end-to-end in the dev vault.
