+++ src/utils/editorInsertion.ts
@@ -3,6 +3,7 @@ import { MarkdownView } from "obsidian";
 import { log } from "../logger/logManager";
 import type { AppendLinkOptions, LinkPlacement } from "../types/linkPlacement";
 import { placementSupportsEmbed } from "../types/linkPlacement";
+import { convertLinkToEmbed } from "./markdownLinks";
 
 /**
  * Returns the active markdown view if it is showing the given file,
@@ -187,58 +188,6 @@ export function insertLinkWithPlacement(
 }
 
 /**
-	* Extracts the target path from a markdown-style link.
-	* Works with both wiki-style and markdown-style links.
-	*/
-function extractMarkdownLinkTarget(link: string): string | null {
-	const markdownPattern = /^\s*!?\[[^\]]*\]\((.+)\)\s*$/;
-	const match = link.match(markdownPattern);
-	if (!match) {
-		return null;
-	}
-
-	let target = match[1].trim();
-	if (!target) {
-		return null;
-	}
-
-	if (target.startsWith("<") && target.endsWith(">")) {
-		target = target.slice(1, -1).trim();
-	}
-
-	return target;
-}
-
-/**
-	* Converts a regular link to an embed by adding the embed prefix (!).
-	* Works with both wiki-style and markdown-style links.
-	*/
-function convertLinkToEmbed(link: string): string {
-	// Embeds must leverage wiki-style transclusions (`![[...]]`) because Obsidian interprets markdown image syntax (`![...](...)`) as attachment images, not note embeds.
-	const trimmed = link.trim();
-	if (trimmed.startsWith("![[") && trimmed.includes("]]")) {
-		return link;
-	}
-
-	const wikiOpenIndex = link.indexOf("[[");
-	const wikiCloseIndex = link.indexOf("]]", wikiOpenIndex + 2);
-	if (wikiOpenIndex !== -1 && wikiCloseIndex !== -1) {
-		const precedingChar = wikiOpenIndex > 0 ? link.charAt(wikiOpenIndex - 1) : "";
-		if (precedingChar === "!") {
-			return link;
-		}
-		return `${link.slice(0, wikiOpenIndex)}!${link.slice(wikiOpenIndex)}`;
-	}
-
-	const markdownTarget = extractMarkdownLinkTarget(link);
-	if (markdownTarget) {
-		return `![[${markdownTarget}]]`;
-	}
-
-	return link.startsWith("!") ? link : `!${link}`;
-}
-
-/**
  * Inserts a link to the specified file into the active view, respecting
  * Obsidian's "New link format" setting.
  *
@@ -283,8 +232,3 @@ export function insertFileLinkToActiveView(
 
 	return true;
 }
-
-export const __test = {
-	convertLinkToEmbed,
-	extractMarkdownLinkTarget,
-} as const;
